requirements.txt
pandas>=1.3.0
matplotlib>=3.4.0
ttkbootstrap>=1.0.0
urllib3>=1.26.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.12.0
tqdm>=4.62.0
tkinterdnd2-universal>=1.0.0
pyinstaller>=5.0.0